" add: test cases to your classes for sprint 2. add: ReminderService class in application package. add: complete MockEmailServer, and EmailService, and after that sprin 3 is complete.
modify: arrange LoanService, Loanrepository, MemberService, and MemberRepository to be more readabile.
modify: rplace getName() method that you were written to search, and ... with my old code: getUserName().

add: adding testcases for domain package (not important but to increase the coverage!).

Note: Im using Mockito to implement sprin 3.

Important Note: when i testes MockEmailServer class, the testcases have passed, but the error has occured! not logical but
it associated with plugins or dependencies in pom.xml, but no problem immediatly.

Conclusion: the project is ready until sprint 3, and we are ready to start with sprint 4.

"